[Identity.Broker] Nest MyClient/MyClientSettings inside snippets class
Per @jsquire feedback on PR #59447: these placeholder types are used only by the documentation snippets, never by tests. Moved them in as private nested classes and deleted the two separate files.
