--- sdk/identity/Azure.Identity.Broker/tests/samples/DocSnippets/MyClient.cs
@@ -1,18 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-namespace Azure.Identity.Broker.Tests.Samples.DocSnippets
-{
-    /// <summary>
-    /// Placeholder client type used by the configuration / DI documentation snippets.
-    /// </summary>
-    internal class MyClient
-    {
-        public MyClient(MyClientSettings settings)
-        {
-            Settings = settings;
-        }
-
-        public MyClientSettings Settings { get; }
-    }
-}
